requirements.txt
# OpsHero Backend Dependencies

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
pyotp==2.9.0

# HTTP Client
httpx==0.26.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# LLM
groq==0.4.2

# Background Tasks (optional, for Celery)
celery==5.3.4

# Monitoring (optional)
prometheus-client==0.19.0

# Utils
python-dateutil==2.8.2
